% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignificanceTestWilcoxon.R
\name{wilcoxon_pairwise}
\alias{wilcoxon_pairwise}
\title{Wilcoxon Test Statistic
Objective :
  To compute pairwise Wilcoxon statistics for a set of predictions
  1. Find all *.F1 files in the given prediction folder
  2. Create a pairwise set for all F1 files
  3. Load F1 file in pairwise manner
  5. Run Wilxocon on the pair
  6. Return Test Statistic and pValue.
  7. Optional: Save results to file, named in pairwise manner.
Usage:
  1. wilcoxon_pairwise(scoresFolder = 'Full folder path')
  2. Results will be saved in *.wil and *.wilpvalue files
Created by: kah siong
Created on: 07/08/17}
\usage{
wilcoxon_pairwise(scoresFolder)
}
\arguments{
\item{A}{full path to the folder containing all the F1 files}
}
\value{
Print out and saved files
}
\description{
Wilcoxon Test Statistic
Objective :
  To compute pairwise Wilcoxon statistics for a set of predictions
  1. Find all *.F1 files in the given prediction folder
  2. Create a pairwise set for all F1 files
  3. Load F1 file in pairwise manner
  5. Run Wilxocon on the pair
  6. Return Test Statistic and pValue.
  7. Optional: Save results to file, named in pairwise manner.
Usage:
  1. wilcoxon_pairwise(scoresFolder = 'Full folder path')
  2. Results will be saved in *.wil and *.wilpvalue files
Created by: kah siong
Created on: 07/08/17
}
\author{
Kah Siong, Tan
}
